requirements.txt
streamlit>=1.35.0
python-dotenv>=1.0.0
pydantic>=2.7.0
pandas>=2.2.0
tqdm>=4.66.0


sentence-transformers>=3.0.0
torch>=2.2.0


huggingface_hub>=0.23.0

pinecone>=3.2.0

fastapi>=0.111.0
uvicorn[standard]>=0.29.0

langsmith>=0.1.75

